requirements.txt
Flask==2.3.3
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
gunicorn==21.2.0
